Refactoring routes test
