+++ controllers/views/creacionController.js
@@ -1,7 +1,9 @@
+const items = require('../../views');
+
 const controller = {
     display: (req,res) => {
        res.render('creacion.ejs')
     }
  }
  
- module.exports = controller;
+module.exports = controller;
