+++ controllers/views/productDetailController.js
@@ -1,4 +1,4 @@
-const items = require('../data/items');
+const items = require('../../views');
 
 const controller = {
    display: (req,res) => {
