+++ controllers/views/registerFormController.js
@@ -1,3 +1,4 @@
+const items = require('../../views');
 
 const controller = {
    display: (req,res) => {
