+++ routes/data/_routes.js
@@ -1,0 +1,5 @@
+const router = require('express').Router();
+
+router.use('/products', require('./products/_routes'));
+
+module.exports = router;
