+++ routes/data/products/_routes.js
@@ -1,0 +1,4 @@
+const router = require('express').Router();
+const productController = require('../../controllers/productController');
+
+router.post('/', productController.crearProducto);
