+++ routes/views/creacion.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const {display} = require('../controllers/loginController');
+const {display} = require('../../controllers/views/creacionController.js');
 
 router.get('/', display);
