+++ routes/views/home.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const homeController = require('../controllers/homeController');
+const homeController = require('../../controllers/views/homeController');
 
 router.get('/', homeController.display);
