+++ routes/views/login.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const {display} = require('../controllers/productDetailController');
+const {display} = require('../../controllers/loginController.js');
 
 router.get('/', display);
