+++ routes/views/productCart.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const {display} = require('../controllers/registerFormController');
+const {display} = require('../../controllers/productCartController.js');
 
 router.get('/', display);
