+++ routes/views/registerForm.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const {display} = require('../controllers/productCartController');
+const {display} = require('../../controllers/registerFormController.js');
 
 router.get('/', display);
