+++ views/home.ejs
@@ -10,7 +10,7 @@
     </header>
   
     <main class="contenedor1">
-            
+        
             <article class="contenedor3">
                 <div id="carrusel">
                     <a href="#" class="left-arrow"><img src="images/left-arrow.png" /></a>
